refactor: implement drawer on Web with CSS transitions
